requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-WTF==1.2.1
python-dotenv==0.19.0
email-validator==2.1.0.post1
Werkzeug==2.0.1
gunicorn==20.1.0
psycopg2-binary==2.9.1
